dac_sine
- Changed I2S interface code to use SD_I2S_INTERFACE directive rather than
hard-coded values.
- Changed receiving of data via I2S to be 32 bit, to accomodate for I2S
standard requirement.
- Set the I2C pins to open drain, to prevent line from being DRIVEN
high.

+++ TLV/inc/ringbuffer.h
@@ -17,16 +17,16 @@
 typedef struct{
 	uint8_t tail;
 	uint8_t head;
-	uint16_t buffer[BUF_SIZE];
+	int32_t buffer[BUF_SIZE];
 } RingBuffer;
 
 void ringBufferInit(RingBuffer*);
 
 // put an element in a buffer
-bool put(RingBuffer*, int16_t);
+bool put(RingBuffer*, int32_t);
 
 // remove an element from the buffer
-int16_t get(RingBuffer*);
+int32_t get(RingBuffer*);
 
 // to check if the buffer is full
 int hasSpace(RingBuffer*);
