+++ TLV/src/ringbuffer.c
@@ -19,7 +19,7 @@ void ringBufferInit(RingBuffer* buffer)
  * @param buffer it is a structure type to access the correct ring buffer.
  * @param element it is the character.
  */
-bool put(RingBuffer* buffer, int16_t element)
+bool put(RingBuffer* buffer, int32_t element)
 {
 	//Check space
 	if (hasSpace(&buffer) != 0)
@@ -38,7 +38,7 @@ bool put(RingBuffer* buffer, int16_t element)
  * Here we have a method to get and remove characters into the buffer.
  * @param buffer it is a structure type to access the correct ring buffer.
  */
-int16_t get(RingBuffer* buffer)
+int32_t get(RingBuffer* buffer)
 {
 	//check space
 	if (hasElement(&buffer) != 0)
